所有接口都对接好了（邢梓奇）

+++ src/main/java/com/springboot/cloud/common/core/entity/vo/BaseVo.java
@@ -8,8 +8,6 @@ import java.util.Date;
 /**
  * @ClassName BaseVo
  * @Description
- * @Author cj
- * @Date 2019/05/30
  */
 public class BaseVo {
 
@@ -23,7 +21,7 @@ public class BaseVo {
     String remark;
 
     @ApiModelProperty(value = "是否删除(0:否|1:是)")
-    Byte isDelete;
+    int isDelete;
 
     @ApiModelProperty(value = "创建人")
     String createdBy;
